retry retrofit requests, include clientRequestId in kato/ops calls
